refactor: Identification Entry Page
