+++ frontend-hybrid/yakal/lib/main.dart
@@ -4,11 +4,11 @@ import 'package:flutter_native_splash/flutter_native_splash.dart';
 import 'package:get/get.dart';
 import 'package:kakao_flutter_sdk_user/kakao_flutter_sdk_user.dart';
 import 'package:yakal/screens/Home/home_screen.dart';
+import 'package:yakal/screens/Login/IdentificationEntry/screen.dart';
 import 'package:yakal/screens/Login/LoginEntry/screen.dart';
 import 'package:yakal/screens/Login/LoginTerms/screen.dart';
 import 'package:yakal/screens/Login/auth_finish_screen.dart';
 import 'package:yakal/screens/Login/enter_nickname_screen.dart';
-import 'package:yakal/screens/Login/identification_entry_screen.dart';
 import 'package:yakal/screens/Login/identification_result_screen.dart';
 import 'package:yakal/screens/Login/identification_screen.dart';
 import 'package:yakal/screens/Login/select_mode_screen.dart';
