+++ frontend-hybrid/yakal/lib/screens/Login/IdentificationEntry/style.dart
@@ -1,0 +1,26 @@
+import 'package:flutter/material.dart';
+
+import '../../../utilities/style/color_styles.dart';
+
+class IdentificationEntryStyle {
+  static const boldTitle = TextStyle(
+    color: ColorStyles.black,
+    fontSize: 24,
+    fontWeight: FontWeight.w700,
+    height: 1.4,
+  );
+
+  static const title = TextStyle(
+    color: ColorStyles.black,
+    fontSize: 24,
+    fontWeight: FontWeight.w500,
+    height: 1.4,
+  );
+
+  static const description = TextStyle(
+    color: ColorStyles.gray5,
+    fontSize: 16,
+    fontWeight: FontWeight.w500,
+    height: 1.6,
+  );
+}
